+++ src/Manipulator/AlfStringManipulator.php
@@ -14,14 +14,22 @@ class AlfStringManipulator extends AlfBasicClass {
         return AlfProgramming::_()->unused($obj, static::_AlfBasicClass($obj));
     }
 
-    public function getStringLength(string $string) : int {
+    public function getStringLength(string $str) : int {
         // without a charset, the length = the byte size.
-        return strlen($string);
+        return strlen($str);
     }
 
-    public function getStringByteSize(string $string) : int {
+    public function getStringByteSize(string $str) : int {
         // without a charset, the length = the byte size.
-        return strlen($string);
+        return strlen($str);
+    }
+
+    public function getStringAsUpperCase(string $str) : string {
+        return strtoupper($str);
+    }
+
+    public function getStringAsLowerCase(string $str) : string {
+        return strtolower($str);
     }
 
 }